// rent-prompts-ai-adding-more-clients/src/payload-types.ts
export interface Media {
  id: string;
  user: string | User;
  alt?: string | null;
  caption?: {
    root: {
      type: string;
      children: {
        type: string;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  } | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
  sizes?: {
    thumbnail?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    card?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    tablet?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
  };
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "users".
 */
export interface User {
  id: string;
  user_name: string;
  coinBalance: number;
  tokens: number;
  role: 'admin' | 'entAdmin' | 'entUser' | 'user';
  products?: (string | Product)[] | null;
  product_files?: (string | ProductFile)[] | null;
  purchases?: (string | Purchase)[] | null;
  rappsPurchases?: (string | RappsPurchase)[] | null;
  bounties?: (string | Bounty)[] | null;
  rapps?: (string | Rapp)[] | null;
  profileImage?: (string | null) | Media;
  coverImage?: (string | null) | Media;
  genInfo?: {
    education?: string | null;
    skills?: string | null;
    gender?: ('male' | 'female' | 'other') | null;
    age?: number | null;
    profession?: string | null;
    workExperience?: number | null;
    interests?:
      | (
          | 'prompt engineering'
          | 'content creation'
          | 'assets'
          | 'ai applications'
          | 'learning and courses'
          | 'blogs'
          | 'bounties'
          | 'community collaboration'
        )[]
      | null;
  };
  socialMediaLinks?: {
    facebook?: string | null;
    instagram?: string | null;
    twitter?: string | null;
    github?: string | null;
    discord?: string | null;
  };
  likes?:
    | {
        user: string | User;
        id?: string | null;
      }[]
    | null;
  following?:
    | {
        user: string | User;
        id?: string | null;
      }[]
    | null;
  followers?:
    | {
        user: string | User;
        id?: string | null;
      }[]
    | null;
  productListing?: number | null;
  userType?: ('individual' | 'enterprise' | 'student') | null;
  over18?: boolean | null;
  domain?: string | null;
  members?: (string | User)[] | null;
  associatedWith?: (string | null) | User;
  rappAccess?:
    | {
        rappId: string | PrivateRapp;
        getAccess?: ('read' | 'delete' | 'create' | 'update')[] | null;
        tokens: number;
        id?: string | null;
      }[]
    | null;
  privateRapps?: (string | PrivateRapp)[] | null;
  createRappPermission?: boolean | null;
  updatedAt: string;
  createdAt: string;
  enableAPIKey?: boolean | null;
  apiKey?: string | null;
  apiKeyIndex?: string | null;
  email: string;
  resetPasswordToken?: string | null;
  resetPasswordExpiration?: string | null;
  salt?: string | null;
  hash?: string | null;
  loginAttempts?: number | null;
  lockUntil?: string | null;
  password?: string | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "products".
 */
export interface Product {
  id: string;
  name: string;
  description: string;
  price: number;
  type: 'text' | 'image' | 'video' | 'music' | 'threed';
  category: 'all' | 'image' | 'text';
  productFiles: string | ProductFile;
  status: 'pending' | 'approved' | 'denied';
  featured: boolean;
  images: {
    image: string | Media;
    id?: string | null;
  }[];
  needsApproval: boolean;
  affiliated: {
    link: string;
    id?: string | null;
  }[];
  user: string | User;
  likes: {
    user: string | User;
    id?: string | null;
  }[];
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "product_files".
 */
export interface ProductFile {
  id: string;
  user: string | User;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "purchases".
 */
export interface Purchase {
  id: string;
  type: 'bounty' | 'assets' | 'course';
  product: string | Product;
  bounty: string | Bounty;
  course: string | Course;
  user: string | User;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "bounties".
 */
export interface Bounty {
  id: string;
  title: string;
  description: {
    root: {
      type: string;
      children: {
        type: string;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  };
  completionDate: string;
  status: 'pending' | 'approved' | 'denied';
  needsApproval: boolean;
  featured: boolean;
  price: number;
  type: 'web' | 'mobile' | 'blockchain' | 'ai' | 'ml';
  attachments: (string | ProductFile)[];
  tags: (
    | 'ai/ml'
    | 'prompt'
    | 'finetuning'
    | 'blockchain'
    | 'frontend'
    | 'backend'
    | 'ui/ux'
    | 'fullstack'
    | 'design'
    | 'testing'
    | 'assets'
    | 'bug'
    | 'chatbot'
  )[];
  user: string | User;
  applicants: {
    user: string | User;
    userName: string;
    linkedin: string;
    phone: number;
    approach: {
      root: {
        type: string;
        children: {
          type: string;
          version: number;
          [k: string]: unknown;
        }[];
        direction: ('ltr' | 'rtl') | null;
        format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
        indent: number;
        version: number;
      };
      [k: string]: unknown;
    };
    relatedFile: (string | ProductFile)[];
    id?: string | null;
  }[];
  createdBy?: string | null;
  updatedBy?: string | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "courses".
 */
export interface Course {
  id: string;
  title: string;
  description: string;
  price: number;
  level: 'beginner' | 'intermediate' | 'advance';
  status: 'pending' | 'approved' | 'denied';
  needsApproval: boolean;
  featured: boolean;
  images: (string | Media)[];
  duration: '1day' | '2days' | '4days' | '1week' | '2week' | '1month' | 'above 1month';
  type: (
    | 'AI Assistants / Agents'
    | 'Advance Prompt Eng'
    | 'Agent Graph'
    | 'RAG'
    | 'Basics of DL'
    | 'Basics of Prompt Eng'
  )[];
  pdf: string | ProductFile;
  user: string | User;
  createdBy?: string | null;
  updatedBy?: string | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "rappsPurchases".
 */
export interface RappsPurchase {
  id: string;
  user: string | User;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "rapps".
 */
export interface Rapp {
  id: string;
  modelType: 'text' | 'image' | 'audio' | 'video';
  model: string | Model;
  key: 'test' | 'prod';
  systemprompt?: string | null;
  prompt?: string | null;
  negativeprompt?: string | null;
  rating: number;
  userRatings?:
    | {
        user: string | User;
        rating: number;
        id?: string | null;
      }[]
    | null;
  userComment?:
    | {
        user: string | User;
        comment: string;
        id?: string | null;
      }[]
    | null;
  status: 'pending' | 'approved' | 'denied';
  slug?: string | null;
  needsApproval: boolean;
  isFeatured: boolean;
  imageinput: boolean;
  purchases?:
    | {
        user: string | User;
        id?: string | null;
      }[]
    | null;
  affiliated_with?: string | null;
  likes?:
    | {
        user: string | User;
        id?: string | null;
      }[]
    | null;
  name: string;
  description: string;
  priceapplicable: boolean;
  price?: number | null;
  totalCost: number;
  getprompt: boolean;
  promptcost?: number | null;
  promptpurchase?:
    | {
        user: string | User;
        id?: string | null;
      }[]
    | null;
  tokens: number;
  images?:
    | {
        image?: (string | null) | Media;
        id?: string | null;
      }[]
    | null;
  creator: string | User;
  systemVariables?:
    | {
        name: string;
        identifier: string;
        displayName: string;
        description: string;
        placeholder?: string | null;
        type: 'string' | 'number' | 'boolean' | 'select';
        allowMultiple: boolean;
        options?: string[] | null;
        id?: string | null;
      }[]
    | null;
  promptVariables?:
    | {
        name: string;
        identifier: string;
        displayName: string;
        description: string;
        placeholder?: string | null;
        type: 'string' | 'number' | 'boolean' | 'select';
        allowMultiple: boolean;
        options?: string[] | null;
        id?: string | null;
      }[]
    | null;
  negativeVariables?:
    | {
        name: string;
        identifier: string;
        displayName: string;
        description: string;
        placeholder?: string | null;
        type: 'string' | 'number' | 'boolean' | 'select';
        allowMultiple: boolean;
        options?: string[] | null;
        id?: string | null;
      }[]
    | null;
  settings?: {
    [k: string]: unknown;
  };
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "models".
 */
export interface Model {
  id: string;
  name: string;
  description: string;
  type: 'text' | 'image' | 'audio' | 'video';
  imageinput: boolean;
  provider: 'groq' | 'openai' | 'replicate' | 'falai';
  settings: {
    name: string;
    type: 'integer' | 'string' | 'boolean' | 'select';
    description: string;
    options?: string[] | null;
    allowMultiple: boolean;
    id?: string | null;
  }[];
  enablePrompt: boolean;
  systemprompt: boolean;
  negativeprompt: boolean;
  prodkeys?: {
    groq?: {
      modelname: string;
      apikey: string;
    };
    replicate?: {
      modelname: string;
      apikey: string;
    };
    openai?: {
      modelname: string;
      apikey: string;
    };
    falai?: {
      modelname: string;
      apikey: string;
    };
  };
  testkeys?: {
    groq?: {
      modelname: string;
      apikey: string;
    };
    replicate?: {
      modelname: string;
      apikey: string;
    };
    openai?: {
      modelname: string;
      apikey: string;
    };
    falai?: {
      modelname: string;
      apikey: string;
    };
  };
  cost: number;
  commissionapplicable: boolean;
  commision?: number | null;
  tokens: number;
  examples: {
    example: string;
    id?: string | null;
  }[];
  updatedAt: string;
  createdAt: string;
}